auth/requirements.txt
fastapi
uvicorn
PyJWT
redis
mysql-connector-python
requests